requirements.txt
anchorpy==0.20.1
anchorpy-core==0.2.0
anyio==3.7.1
apischema==0.16.6
attrs==23.2.0
base58==2.1.1
based58==0.1.1
borsh-construct==0.1.0
cachetools==4.2.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
cryptography==43.0.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupiter-python-sdk==0.0.2.0
more-itertools==8.14.0
OSlash==0.6.3
psycopg2==2.9.9
pycparser==2.22
pyheck==0.1.5
PyNaCl==1.5.0
pysolana==0.2.2
python-dotenv==1.0.1
python-telegram-bot==21.4
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.19.0
sniffio==1.3.1
solana==0.34.2
solders==0.21.0
sumtypes==0.1a6
toml==0.10.2
toolz==0.11.2
types-cachetools==4.2.10
typing_extensions==4.12.2
urllib3==2.2.2
websockets==10.4
